dom/html/test/forms/test_bug1495363.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 1495363</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type="application/javascript">

let readyToCheck = 0;
window.timeoutExpired = () => {
  ok(true, "Timeout expired");
  readyToCheck++;
  maybeCheckResults();
}

window.formSubmitted = () => {
  ok(true, "Form submited!");
  ifr.addEventListener("load", () => {
    readyToCheck++;
    maybeCheckResults();
  }, {once: true});
}

SimpleTest.waitForExplicitFinish();

var ifr = document.createElement('iframe');
ifr.src = "file_bug1495363.html";
document.body.appendChild(ifr);

function maybeCheckResults() {
  ok(readyToCheck <= 2, "So far so good");
  if (readyToCheck < 2) {
    return;
  }

  SimpleTest.executeSoon(() => {
    fetch("file_bug1495363.sjs?result").then(r => r.text()).then(text => {
      let parts = text.split("-");
      is(parts[0], "1", "We have 1 request only");
      is(parts[1], "TIMEOUT", "The request comes from the timer");
      SimpleTest.finish();
    });
  });
}

</script>
</body>
</html>
